pyproject.toml
[project]
name = "popquiz"
version = "0.1.0"
description = "A Django movie quiz application"
requires-python = ">=3.10"
dependencies = [
    "django>=5.2,<6.0",
    "requests",
    "python-dotenv",
    "whitenoise>=6.11.0",
    "django-allauth>=65.14.1",
    "gunicorn>=25.1.0",
    "pillow>=12.1.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = []
